lots of bugs to fix

+++ src/main/java/africa/semicoon/LegendaryHotels/services/ServiceForAdmins.java
@@ -1,35 +1,52 @@
 package africa.semicoon.LegendaryHotels.services;
 
-import africa.semicoon.LegendaryHotels.dto.requests.AdminRequest;
-import africa.semicoon.LegendaryHotels.dto.response.AdminRegistrationResponse;
-import africa.semicoon.LegendaryHotels.dto.response.DeleteResponse;
-import africa.semicoon.LegendaryHotels.dto.response.GetEmailResponse;
+import africa.semicoon.LegendaryHotels.dto.requests.RequestsForAdmins;
+import africa.semicoon.LegendaryHotels.dto.response.ResponseForAdminRegistration;
+import africa.semicoon.LegendaryHotels.dto.response.ResponseForDelete;
+import africa.semicoon.LegendaryHotels.dto.response.ResponseToFindByEmail;
+import africa.semicoon.LegendaryHotels.exceptions.EntityDoesNotExistException;
+import africa.semicoon.LegendaryHotels.exceptions.InvalidEmailException;
 import africa.semicoon.LegendaryHotels.models.Admin;
 import africa.semicoon.LegendaryHotels.repositories.IAdminRepository;
-import africa.semicoon.LegendaryHotels.repositories.RepositoryForAdmin;
+import africa.semicoon.LegendaryHotels.repositories.RepositoryForAdmins;
 import africa.semicoon.LegendaryHotels.utils.Map;
 
 import java.util.List;
 
+import static africa.semicoon.LegendaryHotels.utils.AppUtils.emailIsCorrect;
+
 public class ServiceForAdmins implements IAdminService{
 
-    IAdminRepository adminRepository = new RepositoryForAdmin();
+    IAdminRepository adminRepository = new RepositoryForAdmins();
 
     @Override
-    public AdminRegistrationResponse registerAdmin(AdminRequest adminRequest) {
-        Admin newAdmin = Map.map(adminRequest);
-        adminRepository.saveAdmin(newAdmin);
-        return new AdminRegistrationResponse();
+    public ResponseForAdminRegistration registerAdmin(RequestsForAdmins requestsForAdmins)  throws InvalidEmailException  {
+        if(!emailIsCorrect(requestsForAdmins.getEmail())){
+            throw new InvalidEmailException("Invalid email.");
+        }
+        Admin newAdmin = Map.requestToAdmin(requestsForAdmins);
+        Admin savedAdmin = adminRepository.saveAdmin(newAdmin);
+        return Map.adminToRegistrationResponse(savedAdmin);
     }
 
     @Override
-    public GetEmailResponse findAdminByEmail(String email) {
-        GetEmailResponse getEmailResponse = new GetEmailResponse();
+    public ResponseToFindByEmail findAdminByEmail(RequestsForAdmins requestsForAdmins) throws InvalidEmailException, EntityDoesNotExistException {
+        String email = requestsForAdmins.getEmail();
+        if(!emailIsCorrect(email)){
+            throw new InvalidEmailException("Invalid email.");
+        }
+
+        ResponseToFindByEmail responseToFindByEmail;
         Admin foundAmin = adminRepository.findAdminByEmail(email);
-        getEmailResponse.setFirstName(foundAmin.getFirstName());
-        getEmailResponse.setLastName(foundAmin.getLastName());
-        getEmailResponse.setEmail(foundAmin.getEmail());
-        return getEmailResponse;
+        if(foundAmin == null){
+            throw new EntityDoesNotExistException("Admin does not exist");
+        } else {
+//            getEmailResponse.setFirstName(foundAmin.getFirstName());
+//            getEmailResponse.setLastName(foundAmin.getLastName());
+//            getEmailResponse.setEmail(foundAmin.getEmail());
+            responseToFindByEmail = Map.adminToEmailResponse(foundAmin);
+        }
+        return responseToFindByEmail;
     }
 
     @Override
@@ -38,10 +55,19 @@ public class ServiceForAdmins implements IAdminService{
     }
 
     @Override
-    public DeleteResponse deleteByEmail(String email) {
-        adminRepository.deleteByEmail(email);
-        DeleteResponse deleteResponse = new DeleteResponse();
-        deleteResponse.setMessage("Admin deleted successfully.");
-        return deleteResponse;
+    public ResponseForDelete deleteByEmail(RequestsForAdmins requestsForAdmins) throws InvalidEmailException, EntityDoesNotExistException {
+        if(!emailIsCorrect(requestsForAdmins.getEmail())){
+            throw new InvalidEmailException("Invalid email.");
+        }
+        String email = requestsForAdmins.getEmail();
+
+        if(email == null){
+            throw new EntityDoesNotExistException("Admin does not exist");
+        } else{
+            adminRepository.deleteByEmail(email);
+        }
+        ResponseForDelete responseForDelete = new ResponseForDelete();
+        responseForDelete.setMessage("Admin deleted successfully.");
+        return responseForDelete;
     }
 }
